requirements.txt
langchain
langchain-community
langchainhub
black
python-dotenv
langchain-ollama
